tgosint/requirements.txt
apprise==0.9.9
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
numpy==1.21.6
oauthlib==3.2.0
pandas==1.3.5
pep517==0.12.0
pip-tools==6.6.2
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
smmap==5.0.0
Telethon==1.24.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0